backend/requirements.txt
Django==2.2.2
django-filter==2.1.0
djangorestframework==3.9.3
djangorestframework-simplejwt==4.3.0
dnspython==1.16.0
eventlet==0.24.1
greenlet==0.4.15
gunicorn==19.9.0
isort==4.3.18
lazy-object-proxy==1.4.0
mccabe==0.6.1
meinheld==0.6.1
monotonic==1.5
psycopg2-binary==2.8.2
pycodestyle==2.5.0
PyJWT==1.7.1
pytz==2019.1
rope==0.14.0
six==1.12.0
sqlparse==0.3.0
